Adding the front-end file searching people in database
This UI page will be used to search people with a specific set of skills. The queries sent from here will be captured by the python script to lookup in the database and return a list of matching people. This is being developed standalone now and later Integrated with the main Skill Search application.